Move the processing from snapshot to meadow, and improve it
